More clang-tidy

+++ esphome/components/midi_in/midi.h
@@ -16,7 +16,7 @@ class UARTSerialPort {
  public:
   UARTSerialPort(uart::UARTDevice *uart) { uart_ = uart; }
 
-  void begin(const long baud_rate) {}
+  void begin(const int32 baud_rate) {}
   void end() {}
 
   void write(uint8_t value) { uart_->write(value); };
